fix: scope a bare skip to rewrites and settle lint ranges after the rewrites
- Split bare skip spans out of the off-region vector, so a skip pins layout while its lint diagnostics still report
- Keep an off region suppressing rewrites and lint diagnostics together, which is the only directive covering both
- Collect lint diagnostics once the rewrites settle, so each range resolves against the returned source
- Move `AlignWalker` into its own `walker` module, leaving `Member` and `Settings` in the aligner root
- Derive the reference index's suppression entries from the canonical directive registry, adding the four it omitted
- Correct the alignment family index, three opt-out pointers, and the suppression reference for both behaviors
- Bracket the skipped statement across every line it spans in the scope specimen
