+++ crate/src/primitives/aligner/walker.rs
@@ -1,0 +1,127 @@
+//! The emitter facade the alignment rules drive. Each rule wraps an
+//! `AlignWalker` and calls the `emit_*` methods, which pair the column
+//! math with the skip-hold check and the gap normalization.
+
+use ruff_diagnostics::Edit;
+use ruff_text_size::{TextRange, TextSize};
+
+use super::{
+    Member, Settings, emit::emit_group, is_alignment_candidate, is_held, retain_unheld,
+    space_padding_edit,
+};
+use crate::{rule::RuleId, source::Source};
+
+/// Bundles the `groups` accumulator, `settings`, the owning `rule`, and
+/// borrowed `source` shared by every alignment-rule visitor. Each entry
+/// in `groups` is one fix the pipeline maps to a single diagnostic. The
+/// `rule` id powers the skip-directive check that holds a row out of
+/// its group.
+pub(crate) struct AlignWalker<'a> {
+    pub groups: Vec<Vec<Edit>>,
+    pub rule: RuleId,
+    pub settings: Settings,
+    pub source: &'a Source,
+}
+
+impl<'a> AlignWalker<'a> {
+    /// Builds a walker with an empty `groups` accumulator.
+    pub(crate) fn new(source: &'a Source, settings: Settings, rule: RuleId) -> Self {
+        Self {
+            groups: Vec::new(),
+            rule,
+            settings,
+            source,
+        }
+    }
+
+    /// Aligns `members` to their shared column and folds in a one-space
+    /// rewrite of each gap in `gaps`, recording the combined fix as one
+    /// group. The members-level analog of [`Self::push_with_gaps`],
+    /// pairing the column math of [`Self::group_edits`] with the gap
+    /// normalization.
+    pub(crate) fn emit_group_with_gaps(
+        &mut self,
+        members: &[Member],
+        gaps: impl IntoIterator<Item = TextRange>,
+    ) {
+        let name_edits = self.group_edits(members);
+        self.push_with_gaps(name_edits, gaps);
+    }
+
+    /// Aligns `members` as one fix group when they form an alignment
+    /// candidate, recording nothing otherwise.
+    pub(crate) fn emit_if_candidate(&mut self, members: &[Member]) {
+        self.emit_if_candidate_with_gaps(members, std::iter::empty());
+    }
+
+    /// Aligns `members` as one fix group when they form an alignment
+    /// candidate, folding a one-space rewrite of each gap in `gaps` into
+    /// the same group. Records nothing otherwise. The candidate-gated
+    /// counterpart to [`Self::emit_group_with_gaps`].
+    pub(crate) fn emit_if_candidate_with_gaps(
+        &mut self,
+        members: &[Member],
+        gaps: impl IntoIterator<Item = TextRange>,
+    ) {
+        if is_alignment_candidate(self.source, members) {
+            self.emit_group_with_gaps(members, gaps);
+        }
+    }
+
+    /// Drops the held rows from `members`, then emits the survivors as
+    /// one group when they still form an alignment candidate.
+    pub(crate) fn emit_unheld(&mut self, members: impl IntoIterator<Item = Member>) {
+        let kept = self.retain_unheld(members, |m| m.line_start);
+        self.emit_if_candidate(&kept);
+    }
+
+    /// Computes the alignment edits for `members` without recording
+    /// them, leaving the caller to fold in further edits before
+    /// committing the group through [`Self::push_group`].
+    pub(crate) fn group_edits(&self, members: &[Member]) -> Vec<Edit> {
+        let mut edits = Vec::new();
+        emit_group(self.source, members, self.settings, &mut edits);
+        edits
+    }
+
+    /// Returns `true` when `anchor`'s source line is skip-suppressed for
+    /// this rule.
+    pub(crate) fn is_held(&self, anchor: TextSize) -> bool {
+        is_held(self.source, self.rule, anchor)
+    }
+
+    /// Records `edits` as one fix group, dropping an empty group so a
+    /// no-op pass emits no diagnostic.
+    pub(crate) fn push_group(&mut self, edits: Vec<Edit>) {
+        if !edits.is_empty() {
+            self.groups.push(edits);
+        }
+    }
+
+    /// Records `name_edits` together with a one-space rewrite of each gap
+    /// in `gaps` as one fix group. A gap already holding one space emits
+    /// nothing. The gaps are the secondary spans a rule normalizes beside
+    /// its aligned column, like the `=`-to-value gap or the `:`-to-body
+    /// gap.
+    pub(crate) fn push_with_gaps(
+        &mut self,
+        mut name_edits: Vec<Edit>,
+        gaps: impl IntoIterator<Item = TextRange>,
+    ) {
+        name_edits.extend(
+            gaps.into_iter()
+                .filter_map(|r| space_padding_edit(self.source, r, 1)),
+        );
+        self.push_group(name_edits);
+    }
+
+    /// Returns the rows of `members` whose anchor line is not skip-held
+    /// for this rule. The walker-bound form of the free [`retain_unheld`].
+    pub(crate) fn retain_unheld<M>(
+        &self,
+        members: impl IntoIterator<Item = M>,
+        line_start: impl Fn(&M) -> TextSize,
+    ) -> Vec<M> {
+        retain_unheld(self.source, self.rule, members, line_start)
+    }
+}
