+++ crate/tests/fixtures/suppression/skip_leaves_lint_diagnostics_alone/input.py
