+++ site/.vitepress/theme/components/suppression/ScopeSpecimen.vue
@@ -26,7 +26,9 @@ const lines: SpecimenLine[] = [
   { bracket : 'mid',   scope : 'block', text : '            [0, 0, 0, 1]]' },
   { bracket : 'close', scope : 'block', text : '    # fmt: on' },
   { bracket : 'mid',   scope : 'file',  text : '' },
-  { bracket : 'solo',  scope : 'line',  text : '    out = build(rows)  # fmt: skip' },
+  { bracket : 'open',  scope : 'line',  text : '    out = build(' },
+  { bracket : 'mid',   scope : 'line',  text : '        rows,' },
+  { bracket : 'close', scope : 'line',  text : '    )  # fmt: skip' },
   { bracket : 'solo',  scope : 'line',  text : '    log(out)  # prose: ignore[<rule>]' },
   { bracket : 'mid',   scope : 'file',  text : '' },
   { bracket : 'open',  scope : 'dict',  text : '    STAGES = {  # prose: keep' },
