+++ site/rules/alignment/index.md
@@ -1,6 +1,6 @@
 # Alignment Rules
 
-The alignment rules align a shared token across consecutive rows so columns of assignments, annotations, and patterns read as vertical groups. The shared math lives in the [[aligner]] primitive, and each rule supplies the walker that decides what counts as a group. The per-rule `max-shift` facet bounds how far rows shift to align, wherein each rule walks a run in source order and breaks a fresh group at the first row whose width spread would exceed the cap. `false` lifts the cap so a contiguous run folds into one column, and `0` forbids any shift. A `# prose: skip` on one row holds it out of its group while the rest align around it.
+The alignment rules align a shared token across consecutive rows so columns of assignments, annotations, and patterns read as vertical groups. The shared math lives in the [[aligner]] primitive, and each rule supplies the walker that decides what counts as a group. The per-rule `max-shift` facet bounds how far rows shift to align, wherein each rule walks a run in source order and breaks a fresh group at the first row whose width spread would exceed the cap. `false` lifts the cap so a contiguous run folds into one column, and `0` forbids any shift. A `# prose: skip` holds a row out of its group, so a held single-line row stays transparent and the rest align around it, whereas a held multi-line statement closes the run.
 
 <RuleCardList family="alignment" />
