+++ site/rules/formatting/strip-trailing-commas.md
@@ -22,7 +22,7 @@ The strip is unconditional within the contexts named above, so the rule carries 
 
 <template #related-after>
 
-For per-line opt-outs *(projects that prefer the trailing comma for diff stability even on multi-line forms)*, [**Suppression**](/usage/suppression) covers the `# fmt: off` / `# fmt: on` block markers.
+For block-level opt-outs *(projects that prefer the trailing comma for diff stability even on multi-line forms)*, [**Suppression**](/usage/suppression) covers the `# fmt: off` / `# fmt: on` block markers.
 
 </template>
